feat: Add experimental support for hanging punctuation (overhang)
Backport to resilient of an earlier private experimental feature
(ref: priv-r3exp-prot-stage0) = hanging punctuation support.

This initial version has limitations, see in-code comments.

For the record, further stages considered general "protrusion" (of which
hanging punctuation may be considered as a special case), but the genaral
approaches are not that straightforward, and pose other challenges; while
this simpler use case is doable with less implications.