fix(typesetters): Make bidi work a bit better with hboxes and liners
Stop using the "bidi" package and refactor the core functionality as a
first-class citizen in the silent typesetter.
Move the direction-changing commands from the "bidi" package to our "override"
class.
Try to fix where the bidi logic is invoked so that it works better on hboxes
(horizontal-restricted mode) and in paragraphs with liners (internal reboxing
for multiline effects). This might not be a full fix, but at least some code
paths are simplified and bidi is no longer monkey-hacking methods from the
base typesetter.

Co-authored-by: mimbrown <mimbrown94+git@gmail.com>