feat: Add missing bits from TeX82 line-breaker on adjdemerits and looseness
The looseness missing code is only activated when looseness is not 0,
its default - so won't have any effect with default settings.
The adjdemerits missing code (on fit_class vs current fitness) will
however change some line breakpoints in the general case. It's not
clear why this bit of code was skipped...